+++ lib/src/structs.dart
@@ -1177,8 +1177,55 @@ class OVERLAPPED extends Struct {
 
   external Pointer pointer;
 
+  // Workaround lack of anonymous unions in Dart
+  // (https://github.com/dart-lang/sdk/issues/46501)
+  int get Offset => pointer.address & 0xFFFFFFFF;
+  int get OffsetHigh => (pointer.address >> 32) & 0xFFFFFFFF;
+
+  set Offset(int newValue) {
+    pointer = Pointer.fromAddress(
+        (pointer.address & 0xFFFFFFFF00000000) + (newValue & 0xFFFFFFFF));
+  }
+
+  set OffsetHigh(int newValue) {
+    pointer = Pointer.fromAddress(
+        ((newValue & 0xFFFFFFFF) << 32) + (pointer.address & 0xFFFFFFFF));
+  }
+
   @IntPtr()
   external int hEvent;
+}
+
+// typedef struct _WLAN_RAW_DATA_LIST {
+//     DWORD dwTotalSize;
+//     DWORD dwNumberOfItems;
+//     struct {
+//         // the beginning of the data blob
+//         // the offset is w.r.t. the beginning of the entry
+//         DWORD dwDataOffset;
+//         // size of the data blob
+//         DWORD dwDataSize;
+//     } DataList[1];
+// } WLAN_RAW_DATA_LIST, *PWLAN_RAW_DATA_LIST;
+
+class _DataList extends Struct {
+  @Uint32()
+  external int dwDataOffset;
+  @Uint32()
+  external int dwDataSize;
+}
+
+/// The WLAN_RAW_DATA_LIST structure contains raw data in the form of an
+/// array of data blobs that are used by some Native Wifi functions.
+///
+/// {@category Struct}
+class WLAN_RAW_DATA_LIST extends Struct {
+  @Uint32()
+  external int dwTotalSize;
+  @Uint32()
+  external int dwNumberOfItems;
+  @Array(1)
+  external Array<_DataList> DataList;
 }
 
 // typedef struct mmtime_tag {