+++ lib/src/user32.dart
@@ -1373,6 +1373,27 @@ late final _DragDetect = _user32.lookupFunction<
     Int32 Function(IntPtr hwnd, POINT pt),
     int Function(int hwnd, POINT pt)>('DragDetect');
 
+/// Animates the caption of a window to indicate the opening of an icon or
+/// the minimizing or maximizing of a window.
+///
+/// ```c
+/// BOOL DrawAnimatedRects(
+///   HWND       hwnd,
+///   int        idAni,
+///   const RECT *lprcFrom,
+///   const RECT *lprcTo);
+/// ```
+/// {@category user32}
+int DrawAnimatedRects(
+        int hwnd, int idAni, Pointer<RECT> lprcFrom, Pointer<RECT> lprcTo) =>
+    _DrawAnimatedRects(hwnd, idAni, lprcFrom, lprcTo);
+
+late final _DrawAnimatedRects = _user32.lookupFunction<
+    Int32 Function(
+        IntPtr hwnd, Int32 idAni, Pointer<RECT> lprcFrom, Pointer<RECT> lprcTo),
+    int Function(int hwnd, int idAni, Pointer<RECT> lprcFrom,
+        Pointer<RECT> lprcTo)>('DrawAnimatedRects');
+
 /// The DrawCaption function draws a window caption.
 ///
 /// ```c
@@ -1392,6 +1413,60 @@ late final _DrawCaption = _user32.lookupFunction<
     int Function(
         int hwnd, int hdc, Pointer<RECT> lprect, int flags)>('DrawCaption');
 
+/// The DrawEdge function draws one or more edges of rectangle.
+///
+/// ```c
+/// BOOL DrawEdge(
+///   HDC    hdc,
+///   LPRECT qrc,
+///   UINT   edge,
+///   UINT   grfFlags);
+/// ```
+/// {@category user32}
+int DrawEdge(int hdc, Pointer<RECT> qrc, int edge, int grfFlags) =>
+    _DrawEdge(hdc, qrc, edge, grfFlags);
+
+late final _DrawEdge = _user32.lookupFunction<
+    Int32 Function(IntPtr hdc, Pointer<RECT> qrc, Uint32 edge, Uint32 grfFlags),
+    int Function(
+        int hdc, Pointer<RECT> qrc, int edge, int grfFlags)>('DrawEdge');
+
+/// The DrawFocusRect function draws a rectangle in the style used to
+/// indicate that the rectangle has the focus.
+///
+/// ```c
+/// BOOL DrawFocusRect(
+///   HDC        hDC,
+///   const RECT *lprc);
+/// ```
+/// {@category user32}
+int DrawFocusRect(int hDC, Pointer<RECT> lprc) => _DrawFocusRect(hDC, lprc);
+
+late final _DrawFocusRect = _user32.lookupFunction<
+    Int32 Function(IntPtr hDC, Pointer<RECT> lprc),
+    int Function(int hDC, Pointer<RECT> lprc)>('DrawFocusRect');
+
+/// The DrawFrameControl function draws a frame control of the specified
+/// type and style.
+///
+/// ```c
+/// BOOL DrawFrameControl(
+///   HDC    ,
+///   LPRECT ,
+///   UINT   ,
+///   UINT);
+/// ```
+/// {@category user32}
+int DrawFrameControl(
+        int param0, Pointer<RECT> param1, int param2, int param3) =>
+    _DrawFrameControl(param0, param1, param2, param3);
+
+late final _DrawFrameControl = _user32.lookupFunction<
+    Int32 Function(
+        IntPtr param0, Pointer<RECT> param1, Uint32 param2, Uint32 param3),
+    int Function(int param0, Pointer<RECT> param1, int param2,
+        int param3)>('DrawFrameControl');
+
 /// Draws an icon or cursor into the specified device context.
 ///
 /// ```c
@@ -1408,6 +1483,60 @@ int DrawIcon(int hDC, int X, int Y, int hIcon) => _DrawIcon(hDC, X, Y, hIcon);
 late final _DrawIcon = _user32.lookupFunction<
     Int32 Function(IntPtr hDC, Int32 X, Int32 Y, IntPtr hIcon),
     int Function(int hDC, int X, int Y, int hIcon)>('DrawIcon');
+
+/// The DrawState function displays an image and applies a visual effect to
+/// indicate a state, such as a disabled or default state.
+///
+/// ```c
+/// BOOL DrawStateW(
+///   HDC           hdc,
+///   HBRUSH        hbrFore,
+///   DRAWSTATEPROC qfnCallBack,
+///   LPARAM        lData,
+///   WPARAM        wData,
+///   int           x,
+///   int           y,
+///   int           cx,
+///   int           cy,
+///   UINT          uFlags);
+/// ```
+/// {@category user32}
+int DrawState(
+        int hdc,
+        int hbrFore,
+        Pointer<NativeFunction<DrawStateProc>> qfnCallBack,
+        int lData,
+        int wData,
+        int x,
+        int y,
+        int cx,
+        int cy,
+        int uFlags) =>
+    _DrawState(hdc, hbrFore, qfnCallBack, lData, wData, x, y, cx, cy, uFlags);
+
+late final _DrawState = _user32.lookupFunction<
+    Int32 Function(
+        IntPtr hdc,
+        IntPtr hbrFore,
+        Pointer<NativeFunction<DrawStateProc>> qfnCallBack,
+        IntPtr lData,
+        IntPtr wData,
+        Int32 x,
+        Int32 y,
+        Int32 cx,
+        Int32 cy,
+        Uint32 uFlags),
+    int Function(
+        int hdc,
+        int hbrFore,
+        Pointer<NativeFunction<DrawStateProc>> qfnCallBack,
+        int lData,
+        int wData,
+        int x,
+        int y,
+        int cx,
+        int cy,
+        int uFlags)>('DrawStateW');
 
 /// The DrawText function draws formatted text in the specified rectangle.
 /// It formats the text according to the specified method (expanding tabs,
@@ -3492,6 +3621,22 @@ int GetWindowInfo(int hwnd, Pointer<WINDOWINFO> pwi) =>
 late final _GetWindowInfo = _user32.lookupFunction<
     Int32 Function(IntPtr hwnd, Pointer<WINDOWINFO> pwi),
     int Function(int hwnd, Pointer<WINDOWINFO> pwi)>('GetWindowInfo');
+
+/// Retrieves information about the specified window. The function also
+/// retrieves the value at a specified offset into the extra window memory.
+///
+/// ```c
+/// LONG_PTR GetWindowLongPtrW(
+///   HWND hWnd,
+///   int  nIndex
+/// );
+/// ```
+/// {@category user32}
+int GetWindowLongPtr(int hWnd, int nIndex) => _GetWindowLongPtr(hWnd, nIndex);
+
+late final _GetWindowLongPtr = _user32.lookupFunction<
+    IntPtr Function(IntPtr hWnd, Uint32 nIndex),
+    int Function(int hWnd, int nIndex)>('GetWindowLongPtrW');
 
 /// Retrieves the full path and file name of the module associated with the
 /// specified window handle.
@@ -6451,7 +6596,7 @@ late final _ShowWindow = _user32.lookupFunction<
 int ShowWindowAsync(int hWnd, int nCmdShow) => _ShowWindowAsync(hWnd, nCmdShow);
 
 late final _ShowWindowAsync = _user32.lookupFunction<
-    Int32 Function(IntPtr hWnd, Int32 nCmdShow),
+    Int32 Function(IntPtr hWnd, Uint32 nCmdShow),
     int Function(int hWnd, int nCmdShow)>('ShowWindowAsync');
 
 /// Triggers a visual signal to indicate that a sound is playing.