+++ tool/metadata/projection/win32_typemap.dart
@@ -44,11 +44,14 @@ const win32TypeMap = <String, String>{
   'LPOCNCONNPROCW': 'Pointer<NativeFunction<OpenCardConnProc>>',
   'LPOCNDSCPROC': 'Pointer<NativeFunction<OpenCardDisconnProc>>',
   'TIMERPROC': 'Pointer<NativeFunction<TimerProc>>',
+  'MagImageScalingCallback': 'Pointer<NativeFunction<MagImageScalingCallback>>',
   'PFN_AUTHENTICATION_CALLBACK_EX':
       'Pointer<NativeFunction<PfnAuthenticationCallbackEx>>',
   'PSYM_ENUMERATESYMBOLS_CALLBACKW':
       'Pointer<NativeFunction<SymEnumSymbolsProc>>',
   'LPEXCEPFINO_DEFERRED_FILLIN': 'Pointer<NativeFunction<ExcepInfoProc>>',
+  'WLAN_NOTIFICATION_CALLBACK':
+      'Pointer<NativeFunction<WlanNotificationCallback>>',
 
   // Void pointers
   'LPVOID': 'Pointer', 'PVOID': 'Pointer', 'LPCVOID': 'Pointer',