+++ tool/metadata/winmd_caveats.dart
@@ -15,6 +15,7 @@ const winmdGenerated = [
   'gdi32',
   'kernel32',
   'kernelbase',
+  'magnification',
   'ole32',
   'oleaut32',
   'powrprof',
@@ -29,5 +30,6 @@ const winmdGenerated = [
   'winmm',
   'winscard',
   'winspool',
+  'wlanapi',
   'ws2_32',
 ];